Add NetworkConfig
